% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{BasicKey}
\alias{BasicKey}
\title{Template for Ternary Key}
\usage{
BasicKey(legend_surface, limits, brklab, show_center, center, lwd)
}
\arguments{
\item{legend_surface}{A data frame with numeric 'id', 'p1', 'p2', 'p3' and
character column 'rgb'.}

\item{limits}{A 2 by 3 matrix of lower and upper limits for p1, p2 and p3.}

\item{brklab}{Breaks and labels as returned by \code{\link{BreaksAndLabels}}.}

\item{show_center}{Should the center be marked on the legend? (logical)}

\item{center}{Ternary coordinates of the grey-point.}

\item{lwd}{A numeric scalar giving the linewidth of the legend surface
polygons.}
}
\value{
A ggtern grob.
}
\description{
Return various types of breaks and labels for ternary color keys.
}
\keyword{internal}
