% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{BreaksAndLabels}
\alias{BreaksAndLabels}
\title{Breaks and Labels for Ternary Color Key}
\usage{
BreaksAndLabels(type, center = NULL, breaks = NULL)
}
\arguments{
\item{type}{An integer 1, 2, or 3.}

\item{center}{Ternary coordinates of the grey-point.}

\item{breaks}{Number of breaks in the discrete color scale. An integer >1.
Values above 99 imply no discretization.}
}
\value{
A list of lists containing breaks and labels for each of the 3
  ternary axes.
}
\description{
Return various types of breaks and labels for ternary color keys.
}
\examples{
tricolore:::BreaksAndLabels(1, breaks = 3)
tricolore:::BreaksAndLabels(2)
tricolore:::BreaksAndLabels(3, center = c(1/3, 1/3, 1/3))

}
\keyword{internal}
