% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{DemoTricolore}
\alias{DemoTricolore}
\title{Interactive Tricolore Demonstration}
\usage{
DemoTricolore()
}
\value{
Opens a shiny app session.
}
\description{
An interactive demonstration of the tricolore color scale inspired by the
colorbrewer2.org application. Helps in picking the right color scale for your
data.
}
