% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\docType{data}
\name{euro_education}
\alias{euro_education}
\title{Regional Composition of Educational Attainment in Europe 2016}
\format{A data frame with 324 rows and 4 variables:
  \describe{
    \item{id}{NUTS-2 regional code.}
    \item{ed0_2}{Share of population with highest attained education "lower secondary or less".}
    \item{ed3_4}{Share of population with highest attained education "upper secondary".}
    \item{ed5_8}{Share of population with highest attained education "tertiary".}
  }}
\source{
Derived from Eurostats table "edat_lfse_04".
}
\usage{
euro_education
}
\description{
A dataset containing the relative share of population by educational
attainment in the European NUTS-2 regions 2016.
}
\details{
Covers ages 25 to 64.
}
\keyword{datasets}
