% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\docType{data}
\name{euro_geo_nuts2}
\alias{euro_geo_nuts2}
\title{NUTS-2 Level Polygons for Europe}
\format{A data frame with 4,206 rows and 7 variables:
  \describe{
    \item{long}{Longitude.}
    \item{lat}{Latitude.}
    \item{order}{Drawing order of polygon path.}
    \item{hole}{Features hole?}
    \item{piece}{Piece id.}
    \item{id}{NUTS-2 code.}
    \item{group}{Drawing group.}
  }}
\source{
Derived from Eurostats European Geodata.
  (c) EuroGeographics for the administrative boundaries.
  \url{http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/}
}
\usage{
euro_geo_nuts2
}
\description{
A dataset containing the NUTS-2 level polygons of Europes regions.
}
\keyword{datasets}
