% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\docType{data}
\name{euro_sectors}
\alias{euro_sectors}
\title{Regional Labour Force Composition in Europe 2008 to 2016}
\format{A data frame with 2,902 rows and 5 variables:
  \describe{
    \item{year}{Year.}
    \item{id}{NUTS-2 regional code.}
    \item{primary}{Share of labour force in primary sector.}
    \item{secondary}{Share of labour force in secondary sector.}
    \item{tertiary}{Share of labour force in tertiary sector.}
  }}
\source{
Derived from Eurostats table "lfst_r_lfe2en2".
}
\usage{
euro_sectors
}
\description{
A dataset containing the relative share of workers by labour force sector
in the European NUTS-2 regions from 2008 to 2016.
}
\details{
The original NACE (rev. 2) codes have been recoded into the three sectors
  "primary" (A), "secondary" (B-E & F) and "tertiary" (all other NACE codes).
}
\keyword{datasets}
