% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{GeometricMean}
\alias{GeometricMean}
\title{Geometric Mean}
\usage{
GeometricMean(x, na.rm = TRUE, zero.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Should NAs be removed? (default=TRUE)}

\item{zero.rm}{Should zeros be removed? (default=TRUE)}
}
\value{
The geometric mean as numeric scalar.
}
\description{
Calculate the geometric mean for a numeric vector.
}
\examples{
# NOTE: only intended for internal use and not part of the API
tricolore:::GeometricMean(0:100)
tricolore:::GeometricMean(0:100, zero.rm = FALSE)

}
\keyword{internal}
