% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{GetMixture}
\alias{GetMixture}
\title{RGB Mixture of Ternary Composition}
\usage{
GetMixture(P, k, h_, c_, l_, contrast, center)
}
\arguments{
\item{P}{n by 3 matrix of ternary compositions {p1, p2, p3}_i for
i=1, ..., n.}

\item{k}{Number of breaks in the discrete color scale. An integer >0.
Values above 99 imply no discretization.}

\item{h_}{Primary hue of the first ternary element in angular degrees [0, 360).}

\item{c_}{Maximum possible chroma of mixed colors >= 0.}

\item{l_}{Lightness of mixed colours [0, 100].}

\item{center}{Should the composition be centered? (default=TRUE)}
}
\value{
A vector of rgbs hex-codes representing the mixed colors.
}
\description{
Return the ternary balance scheme colors for a matrix of ternary compositions.
}
\examples{
P <- prop.table(matrix(runif(9), ncol = 3), 1)
GetMixture(P, k = 5, h_ = 80, c_ = 170, l_ = 80, contrast = 0.6, center = TRUE)

}
\keyword{internal}
