% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{GetNearest}
\alias{GetNearest}
\title{For Ternary Coordinates P Return the Nearest Coordinate in Set C}
\usage{
GetNearest(P, C, index = FALSE)
}
\arguments{
\item{P, C}{n by 3 matrix of ternary coordinates {p1, p2, p3}_i for
i=1,...,n. n may be different for P and C.}
}
\value{
A matrix of ternary coordinates in C.
}
\description{
For Ternary Coordinates P Return the Nearest Coordinate in Set C
}
\examples{
P <- prop.table(matrix(runif(9), ncol = 3), 1)
C <- tricolore:::GetCentroids(2)[,-1]
tricolore:::GetNearest(P, C)
}
