% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{MixColor}
\alias{MixColor}
\title{Ternary Balance Scheme Colors}
\usage{
MixColor(df, p1, p2, p3, k = Inf, hue = 0.3, chroma = 0.8,
  lightness = 0.8, contrast = 0.6, center = FALSE, legend = FALSE,
  color_space = "hcl")
}
\arguments{
\item{df}{Data frame.}

\item{p1}{Unquoted column name for variable in df giving first proportion
of ternary composition.}

\item{p2}{Unquoted column name for variable in df giving second proportion
of ternary composition.}

\item{p3}{Unquoted column name for variable in df giving third proportion
of ternary composition.}

\item{hue}{Primary hue of the first ternary element [0, 1].}

\item{chroma}{Maximum possible chroma of mixed colors [0, 1].}

\item{lightness}{Lightness of mixed colours [0, 1].}

\item{center}{Should the composition be centered? (default=TRUE)}

\item{legend}{Should a legend be returned along with the colors? (default=FALSE)}
}
\value{
legend=FALSE: A vector of rgbs hex-codes representing the ternary
        balance scheme colors. legend=TRUE: A list with elements "hexsrgb"
        and "legend".
}
\description{
Return the ternary balance scheme colors for ternary compositions.
}
\examples{
P <- as.data.frame(prop.table(matrix(runif(3^6), ncol = 3), 1))
MixColor(P, V1, V2, V3, legend = TRUE)

}
