% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{PlotLegend}
\alias{PlotLegend}
\title{Ternary Balance Scheme Legend}
\usage{
PlotLegend(k, hue = 80, chroma = 170, lightness = 80, contrast = 0.6,
  center = FALSE)
}
\arguments{
\item{k}{Number of breaks on each axis.}

\item{hue}{Primary hue of the first ternary element in angular degrees [0, 360).}

\item{chroma}{Maximum possible chroma of mixed colors >= 0.}

\item{lightness}{Lightness of mixed colours [0, 100].}

\item{center}{Should the composition be centered? (default=TRUE)}
}
\description{
Plot a ternary balance scheme legend.
}
