% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{ScaleVariance}
\alias{ScaleVariance}
\title{Scale Composition To Unity Variance}
\usage{
ScaleVariance(P)
}
\arguments{
\item{P}{n by m matrix of compositions {p1, ..., pm}_i for
i=1,...,n.}
}
\value{
An n by m matrix of compositions with variance scaled to unity.
}
\description{
Scale a compositional data set to unity variance.
}
\examples{
P <- prop.table(matrix(runif(300), 100), margin = 1)
ScaleVariance(P)
tricolore:::TotalVariance(P)
tricolore:::TotalVariance(ScaleVariance(P))

}
\references{
Pawlowsky-Glahn, V., Egozcue, J. J., & Tolosana-Delgado, R. (2007).
  Lecture Notes on Compositional Data Analysis.
}
\keyword{internal}
