% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{TernaryCenterGrid}
\alias{TernaryCenterGrid}
\title{Return Ternary Gridlines Centered Around Some Composition}
\usage{
TernaryCenterGrid(center, spacing)
}
\arguments{
\item{center}{The center of the grid.
A vector of ternary coordinates [p1, p2, p3].}

\item{spacing}{The spacing of the grid in percent-point increments.
A numeric > 0.}
}
\value{
A list of lists.
}
\description{
Return Ternary Gridlines Centered Around Some Composition
}
\examples{
# NOTE: only intended for internal use and not part of the API
tricolore:::TernaryCenterGrid(c(1/6, 2/6, 3/6), 10)

}
\keyword{internal}
