% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{TernaryLimits}
\alias{TernaryLimits}
\title{Return the Limits of Ternary Coordinates}
\usage{
TernaryLimits(P, na.rm = TRUE)
}
\arguments{
\item{P}{n by 3 matrix of ternary coordinates [p1, p2, p3](i) for
i=1,...,n.}

\item{na.rm}{Should NAs be removed? (default=TRUE)}
}
\value{
A 2 by 3 matrix of lower and upper limits for p1, p2 and p3.
}
\description{
Return the Limits of Ternary Coordinates
}
\examples{
# NOTE: only intended for internal use and not part of the API
P <- prop.table(matrix(runif(9), ncol = 3), 1)
tricolore:::TernaryLimits(P)

}
\keyword{internal}
