% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{TotalVariance}
\alias{TotalVariance}
\title{Total Variance of Composition}
\usage{
TotalVariance(P, na.rm = TRUE, zero.rm = TRUE)
}
\arguments{
\item{P}{n by m matrix of compositions {p1, ..., pm}_i for
i=1,...,n.}
}
\value{
The total variance of the compositional data set as a scalar.
}
\description{
Compute the total variance of a compositional data set.
}
\examples{
P <- prop.table(matrix(runif(300), 100), margin = 1)
tricolore:::TotalVariance(P)
P2 <- rbind(P, c(0, 0, 1))
tricolore:::TotalVariance(P2)
tricolore:::TotalVariance(P2, zero.rm = FALSE)

}
\references{
Pawlowsky-Glahn, V., Egozcue, J. J., & Tolosana-Delgado, R. (2007).
  Lecture Notes on Compositional Data Analysis.
}
\keyword{internal}
