% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{ValidateMainArguments}
\alias{ValidateMainArguments}
\title{Validate Main Arguments of tricolore Function}
\usage{
ValidateMainArguments(df, p1, p2, p3)
}
\arguments{
\item{df}{Data frame.}

\item{p1}{Column name for variable in df giving first proportion
of ternary composition (string).}

\item{p2}{Column name for variable in df giving second proportion
of ternary composition (string.}

\item{p3}{Column name for variable in df giving third proportion
of ternary composition (string).}
}
\description{
Validate Main Arguments of tricolore Function
}
\keyword{internal}
