% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{ValidateParameters}
\alias{ValidateParameters}
\title{Validate Parameters}
\usage{
ValidateParameters(pars)
}
\arguments{
\item{pars}{List of parameters.}
}
\description{
Validate parameters of tricolore function.
}
\keyword{internal}
