% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{ValidateParametersShared}
\alias{ValidateParametersShared}
\title{Validate Shared Parameters}
\usage{
ValidateParametersShared(pars)
}
\arguments{
\item{pars}{A named list of parameters.}
}
\description{
Validate parameters shared across tricolore functions.
}
\keyword{internal}
