% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{ValidateParametersTricolore}
\alias{ValidateParametersTricolore}
\title{Validate Tricolore Parameters}
\usage{
ValidateParametersTricolore(pars)
}
\arguments{
\item{pars}{A named list of parameters.}
}
\description{
Validate parameters of Tricolore function.
}
\keyword{internal}
