% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{ValidateParametersTricoloreSextant}
\alias{ValidateParametersTricoloreSextant}
\title{Validate TricoloreSextant Parameters}
\usage{
ValidateParametersTricoloreSextant(pars)
}
\arguments{
\item{pars}{A named list of parameters.}
}
\description{
Validate parameters of TricoloreSextant function.
}
\keyword{internal}
