% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{Centre}
\alias{Centre}
\title{Centre Composition}
\usage{
Centre(P)
}
\arguments{
\item{P}{n by m matrix of compositions {p1, ..., pm}_i for
i=1,...,n.}
}
\value{
n by m matrix of centred compositions.
}
\description{
Centre a compositional data set.
}
\examples{
P <- prop.table(matrix(runif(300), 100), margin = 1)
Centre(P)

}
\references{
Von Eynatten, H., Pawlowsky-Glahn, V., & Egozcue, J. J. (2002).
Understanding perturbation on the simplex: A simple method to better
visualize and interpret compositional data in ternary diagrams.
Mathematical Geology, 34(3), 249-257.
}
\keyword{internal}
