% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\docType{data}
\name{euro_basemap}
\alias{euro_basemap}
\title{Flat Map of European Continent}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\source{
Derived from Eurostats European Geodata.
  (c) EuroGeographics for the administrative boundaries.
  \url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/}
}
\usage{
euro_basemap
}
\description{
A ggplot object rendering a flat background map of the European continent.
}
\keyword{datasets}
