% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\docType{data}
\name{euro_basemap}
\alias{euro_basemap}
\title{Flat Map of European Continent}
\format{
An object of class \code{ggplot} (inherits from \code{ggplot2::ggplot}, \code{ggplot2::gg}, \code{S7_object}, \code{gg}) of length 1.
}
\source{
Derived from geodata provided by the Natural Earth project.
  \url{https://www.naturalearthdata.com/}
}
\usage{
euro_basemap
}
\description{
A ggplot object rendering a flat background map of the European continent.
}
\keyword{datasets}
