% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricolore.R
\name{PlotLegend}
\alias{PlotLegend}
\title{Ternary Balance Scheme Legend}
\usage{
PlotLegend(k, hue = 0.3, chroma = 0.8, lightness = 0.8, contrast = 0.6,
  center = FALSE, color_space = "hcl")
}
\arguments{
\item{k}{Number of breaks on each axis.}

\item{hue}{Primary hue of the first ternary element [0, 1].}

\item{chroma}{Maximum possible chroma of mixed colors [0, 1].}

\item{lightness}{Lightness of mixed colours [0, 1].}

\item{center}{Should the composition be centered? (default=TRUE)}
}
\description{
Plot a ternary balance scheme legend.
}
